+++ packages/components/src/search-control/index.tsx
@@ -2,6 +2,7 @@
  * External dependencies
  */
 import clsx from 'clsx';
+import type { ForwardedRef } from 'react';
 
 /**
  * WordPress dependencies
@@ -16,11 +17,12 @@ import deprecated from '@wordpress/deprecated';
  * Internal dependencies
  */
 import Button from '../button';
+import { InputControlPrefixWrapper } from '../input-control/input-prefix-wrapper';
+import { InputControlSuffixWrapper } from '../input-control/input-suffix-wrapper';
 import type { WordPressComponentProps } from '../context/wordpress-component';
 import type { SearchControlProps, SuffixItemProps } from './types';
-import type { ForwardedRef } from 'react';
 import { ContextSystemProvider } from '../context';
-import { StyledInputControl, SuffixItemWrapper } from './styles';
+import { StyledInputControl } from './styles';
 
 function SuffixItem( {
 	searchRef,
@@ -29,7 +31,7 @@ function SuffixItem( {
 	onClose,
 }: SuffixItemProps ) {
 	if ( ! onClose && ! value ) {
-		return <Icon icon={ search } />;
+		return null;
 	}
 
 	if ( onClose ) {
@@ -44,12 +46,14 @@ function SuffixItem( {
 	};
 
 	return (
-		<Button
-			size="small"
-			icon={ closeSmall }
-			label={ onClose ? __( 'Close search' ) : __( 'Reset search' ) }
-			onClick={ onClose ?? onReset }
-		/>
+		<InputControlSuffixWrapper variant="control">
+			<Button
+				size="small"
+				icon={ closeSmall }
+				label={ onClose ? __( 'Close search' ) : __( 'Reset search' ) }
+				onClick={ onClose ?? onReset }
+			/>
+		</InputControlSuffixWrapper>
 	);
 }
 
@@ -114,15 +118,18 @@ function UnforwardedSearchControl(
 				autoComplete="off"
 				placeholder={ placeholder }
 				value={ value ?? '' }
+				prefix={
+					<InputControlPrefixWrapper variant="icon">
+						<Icon icon={ search } fill="currentColor" />
+					</InputControlPrefixWrapper>
+				}
 				suffix={
-					<SuffixItemWrapper size={ size }>
-						<SuffixItem
-							searchRef={ searchRef }
-							value={ value }
-							onChange={ onChange }
-							onClose={ onClose }
-						/>
-					</SuffixItemWrapper>
+					<SuffixItem
+						searchRef={ searchRef }
+						value={ value }
+						onChange={ onChange }
+						onClose={ onClose }
+					/>
 				}
 				{ ...filteredRestProps }
 			/>