+++ packages/components/CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   `SearchControl`: Move search icon to prefix position ([#71984](https://github.com/WordPress/gutenberg/pull/71984)).
+
 ## 30.5.0 (2025-10-01)
 
 ### Enhancements
 
 -   `TextareaControl`: Add default resize: vertical rule ([#71736](https://github.com/WordPress/gutenberg/pull/71736)).
--   `SearchControl`: Move search icon to prefix position ([#71984](https://github.com/WordPress/gutenberg/pull/71984)).
 
 ### Bug Fixes