+++ packages/components/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Enhancements
 
+-   `SearchControl`: Move search icon to prefix position ([#71984](https://github.com/WordPress/gutenberg/pull/71984)).
+
+## 30.5.0 (2025-10-01)
+
+### Enhancements
+
 -   `TextareaControl`: Add default resize: vertical rule ([#71736](https://github.com/WordPress/gutenberg/pull/71736)).
 
 ### Bug Fixes